+++ src/lib/components/FormData/config.ts
@@ -1,0 +1,5 @@
+import { ConfigInterface } from "./interfaces";
+
+const config: ConfigInterface = {};
+
+export default config;